refactor(checker): route function return branch relation guard
